moving config to main class
