+++ src/main/java/com/tramonti/weather/WeatherApplication.java
@@ -2,21 +2,36 @@ package com.tramonti.weather;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Profile;
+import org.springframework.context.annotation.*;
+import org.springframework.web.client.RestTemplate;
+
+import javax.validation.Validation;
+import javax.validation.Validator;
+import javax.validation.ValidatorFactory;
 
 @SpringBootApplication
 public class WeatherApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(WeatherApplication.class, args);
-	}
+    public static void main(String[] args) {
+        SpringApplication.run(WeatherApplication.class, args);
+    }
 
-	@Configuration
-	@Profile("local")
-	@ComponentScan(lazyInit = true)
-	static class LocalConfig{
-//		-Dspring.profiles.active=local
-	}
+    @Bean
+    @Scope("prototype")
+    public RestTemplate getRestTemplate() {
+        return new RestTemplate();
+    }
+
+    @Bean
+    public Validator getValidator() {
+        ValidatorFactory factory = Validation.buildDefaultValidatorFactory();
+        return factory.getValidator();
+    }
+
+    @Configuration
+    @Profile("dev")
+    @ComponentScan(lazyInit = true)
+    static class LocalConfig {
+//		-Dspring.profiles.active=dev
+    }
 }
