--- src/main/java/com/tramonti/weather/WeatherConfiguration.java
@@ -1,26 +1,0 @@
-package com.tramonti.weather;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Scope;
-import org.springframework.web.client.RestTemplate;
-
-import javax.validation.Validation;
-import javax.validation.Validator;
-import javax.validation.ValidatorFactory;
-
-@Configuration
-public class WeatherConfiguration {
-
-    @Bean
-    @Scope("prototype")
-    public RestTemplate getRestTemplate() {
-        return new RestTemplate();
-    }
-
-    @Bean
-    public Validator getValidator() {
-        ValidatorFactory factory = Validation.buildDefaultValidatorFactory();
-        return factory.getValidator();
-    }
-}
